requirements.txt
Django==2.0.1
psycopg2==2.7.3.2
python-decouple==3.1

pkg-resources==0.0.0
pytz==2017.3


django-debug-toolbar==1.9.1

Pillow==5.0.0

django-solo==1.1.3
celery[redis]==4.1.0
